src/pyproject.toml
[tool.poetry]
name = "your-project-name"
version = "0.1.0"
description = ""
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = ">=0.100.0"
uvicorn = ">=0.20.0"
python-dotenv = ">=1.0.0"
pydantic = ">=2.0.0"
langchain = ">=0.0.300"
langchain-core = ">=0.1.0"
langchain-community = ">=0.2.17"
langchain-openai = ">=0.1.25"
langchain-faiss = ">=0.0.1"
langchain-milvus = ">=0.1.5"
langchain-ollama = ">=0.1.3"
openai = ">=1.52.0"
tiktoken = ">=0.7.0"
faiss-cpu = ">=1.7.4"
sqlalchemy = ">=2.0.0"
rank-bm25 = ">=0.2.2"
numpy = ">=1.24.0"
pandas = ">=2.0.0"
python-multipart = ">=0.0.6"
beautifulsoup4 = ">=4.12.3"
streamlit = ">=1.39.0"
requests = ">=2.31.0"
tqdm = ">=4.65.0"
PyYAML = ">=6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"